Version 2.5.25:
Support RF cavities in DYNK IF they are active (i.e. switched on in fort.2 -> kp(ix) == 6)
Supported attributes:
- voltage
- harmonic
- lag_angle

General bugfixes:
- In daten, always set kp(ix) = 0 for disabled cavities, even if an enabled cavity has not yet been seen
- Fix kz = -12 in trauthin

Note: The bugfixes changes the result of any lattice which has a disabled cavity present in the SINGLE ELEMENTS list, and this disabled cavity is not listed after an active cavity. This includes the usual fort.2s used for the LHC (200 MHz RF system).