Bugfix release 2.5.27:
(1) Bugfix in the function ran_gauss, which is used by collimation:
This function was not initialized correctly, meaning that it might have returned wrong results on the first call (depending on the compiler and the state of the computer).
The first call would most likely be when generating the phase-space position of the first particle in the distribution. This is now fixed.

(2) Increased input checking in DYNK:
There were a few places where DYNK was susceptible to buffer overflow if the given filenames or function names were too long (longer than 20 characters). This could lead to wrong results or program crash.
The problem is now fixed, as the functions and input parser are now treating the string data more carefully.

--- Kyrre