Fix linking of NAFF to LibFFTW, using a find_package from EIGEN.
Note that CMake suddenly started to autodetect SixTrack as a C++ executable and tried to link it as such; this is avoided by explicitly specifying the target language.
Such linking triggered the use of the gcc flags -rdynamic and -Wl,-Bdynamic (together with -static), which caused the executable to not run.
When trying to run, this resulted in the error message "/lib/ld64.so.1: bad ELF interpreter: No such file or directory".