bug fix in online aperture checking
The bug consisted in aperture_saveLastCoordinates called with iBack=-1
at every entry in the lattice structure, leading to the backtracking
algorithm to be skipped most of the times. Now subroutine is called
only if at least a particle is really lost; call moved inside lostpart,
right after the array compaction.