making brho a parameter in mod_common module in source/common_modules.f90:    * brho checkpointed/restared;    * brho updated in part_updateRefEnergy subroutine in      source/mod_particles.f90;    * removed the same stuff from hions module - not really used...    * elens module modified accordingly;
At the same time, found a bug in the calculation of brho, i.e. e0 was used
instead of e0f); this triggered the need to re-set all eLens tests. At the
same time:
   * plotting scripts modified to better lable R2 and give a more human-readable
     error message when positional coordinates are changed;
   * made elensidealthck6d test more sensible:
      * eLens with radial profile extends between 3 and 10 sig, instead of
        between 5 and 10 -> beginning of radial part more visible;
      * powering of 3rd eLens adjusted to match the expected theta_R2;
   * made elensidealthin6d_ramp_DYNK more sensible:
      * reset elens powering to match expected theta_R2
      * shrinked radial profile, to have all particles surviving

At the same time, elens module modified also to for a more reasonable code
factoring when (re-)computing theta_R2 of elenses - in particular elensThetas
subroutine gone, and elens_geo_norm is reported as well on sixtrack STDOUT;